chore: restructure tests
